+++ examples/src/virtual/react/list/material-ui/src/code.tsx
@@ -1,13 +1,13 @@
 import { memo } from "react";
-import { useVirtual, List } from "@af-utils/virtual-react";
+import { useVirtual, List, createListItemRef } from "@af-utils/virtual-react";
 import MuiListNative from "@mui/material/List";
 import ListItem from "@mui/material/ListItem";
 import ListItemButton from "@mui/material/ListItemButton";
 import ListItemText from "@mui/material/ListItemText";
 import type { ListItemProps } from "@af-utils/virtual-react";
 
-const Item = memo<ListItemProps>(({ i, model }) => (
-    <ListItem ref={el => model.el(i, el)} disablePadding>
+const Item = memo<ListItemProps>(({ model, i }) => (
+    <ListItem ref={createListItemRef(model, i)} disablePadding>
         <ListItemButton>
             <ListItemText primary={`row ${i}`} />
         </ListItemButton>