+++ examples/src/virtual/react/list/prepend-items/src/code.tsx
@@ -1,12 +1,16 @@
 import { memo, useRef, useState } from "react";
-import { useVirtualModel, List } from "@af-utils/virtual-react";
+import {
+    useVirtualModel,
+    List,
+    createListItemRef
+} from "@af-utils/virtual-react";
 import { randNumber, randFullName } from "@ngneat/falso";
 import type { ListItemProps } from "@af-utils/virtual-react";
 import css from "./style.module.css";
 
-const Item = memo<ListItemProps>(({ i, model, data }) => (
+const Item = memo<ListItemProps>(({ model, i, data }) => (
     <div
-        ref={el => model.el(i, el)}
+        ref={createListItemRef(model, i)}
         className={css.item}
         style={{ padding: `${data[i].height}px 0.5em` }}
     >