+++ examples/src/virtual/react/list/scroll-to-item/src/code.tsx
@@ -1,5 +1,5 @@
 import { useState, memo, useEffect } from "react";
-import { useVirtual, List } from "@af-utils/virtual-react";
+import { useVirtual, List, createListItemRef } from "@af-utils/virtual-react";
 import { VirtualScroller } from "@af-utils/virtual-core";
 import type { ListItemProps } from "@af-utils/virtual-react";
 import type { FormEvent } from "react";
@@ -7,9 +7,9 @@ import css from "./style.module.css";
 
 const DEFAULT_ROW_COUNT = 50000;
 
-const Item = memo<ListItemProps>(({ i, model, data: pseudoRandomSizes }) => (
+const Item = memo<ListItemProps>(({ model, i, data: pseudoRandomSizes }) => (
     <div
-        ref={el => model.el(i, el)}
+        ref={createListItemRef(model, i)}
         className={css.item}
         style={{
             padding: `${pseudoRandomSizes[i]}px 0.7em`