+++ examples/src/virtual/react/list/sticky-header-and-footer/src/code.tsx
@@ -1,10 +1,10 @@
 import { memo } from "react";
-import { useVirtual, List } from "@af-utils/virtual-react";
+import { useVirtual, List, createListItemRef } from "@af-utils/virtual-react";
 import type { ListItemProps } from "@af-utils/virtual-react";
 import css from "./style.module.css";
 
-const Item = memo<ListItemProps>(({ i, model }) => (
-    <div ref={el => model.el(i, el)} className={css.item}>
+const Item = memo<ListItemProps>(({ model, i }) => (
+    <div ref={createListItemRef(model, i)} className={css.item}>
         row {i}
     </div>
 ));