+++ examples/src/virtual/react/list/load-on-demand/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @af-utils/virtual-examples-react-list-load-on-demand
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @af-utils/virtual-react@1.0.0
+  - @af-utils/virtual-core@1.0.0
+
 ## 0.0.8
 
 ### Patch Changes