+++ website/CHANGELOG.md
@@ -1,5 +1,13 @@
 # website
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @af-utils/virtual-react@1.0.0
+  - @af-utils/virtual-core@1.0.0
+
 ## 0.0.4
 
 ### Patch Changes